fix call to get_title and ytdl_download with wrong arguments
In the case the user_input is not a url, `ytdl_download` was called with
only a video id. This commit fixes this by forming a fully qualified
video URL from the looked-up video id, and using the result as argument
to `ytdl_download`.
Another addressed issue is a typo where `get_title` is called with
`song` instead of `song_url` as argument.